requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-simple-menu==2.1.3
gpiocontrol==1.1
sqlparse==0.4.4
